+++ app/[lang]/components/layouts/HeroImage.tsx
@@ -34,7 +34,7 @@ const HeroImage = ({
     <div className="mt-9 h-full w-full">
       <div className="relative left-1/2 h-[303.552px] w-[458.30400000000003px] -translate-x-1/2 overflow-hidden xs:h-96 xs:w-full">
         <Image
-          src={photo ? photo.url : "/hero_image.jpg"}
+          src={photo ? photo.url : "../hero_image.jpg"}
           alt="Loby picture"
           fill
           sizes="100vw"
