+++ app/[lang]/loading.tsx
@@ -1,6 +1,6 @@
 export default function Loading() {
   return (
-    <div className="fixed top-0 z-[60] flex h-screen w-full items-center justify-center scroll-smooth bg-cdd-red/40">
+    <div className="absolute top-0 z-[60] flex h-screen w-full items-center justify-center scroll-smooth bg-cdd-red/40">
       <h1 className="rounded-full bg-white px-5 py-2">Loading...</h1>
     </div>
   );
