+++ package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@formatjs/intl-localematcher": "^0.5.2",
     "@prisma/client": "^5.7.0",
-    "@react-google-maps/api": "^2.19.2",
+    "@react-google-maps/api": "^2.19.3",
     "@types/node": "20.5.1",
     "@types/react": "18.2.20",
     "@types/react-dom": "18.2.7",
@@ -25,7 +25,7 @@
     "mongodb": "^5.8.0",
     "mongoose": "^7.4.4",
     "negotiator": "^0.6.3",
-    "next": "^13.5.6",
+    "next": "^14.1",
     "postcss": "^8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
