+++ public/en/hero_image.jpg
