+++ public/en/site.webmanifest
@@ -1,14 +1,14 @@
 {
-  "name": "",
-  "short_name": "",
+  "name": "Casona diez diez",
+  "short_name": "Casona diez",
   "icons": [
     {
-      "src": "/android-chrome-192x192.png",
+      "src": "/../android-chrome-192x192.png",
       "sizes": "192x192",
       "type": "image/png"
     },
     {
-      "src": "/android-chrome-512x512.png",
+      "src": "/../android-chrome-512x512.png",
       "sizes": "512x512",
       "type": "image/png"
     }
