+++ public/es/site.webmanifest
@@ -1,0 +1,19 @@
+{
+  "name": "Casona diez diez",
+  "short_name": "Casona diez",
+  "icons": [
+    {
+      "src": "/../android-chrome-192x192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/../android-chrome-512x512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ],
+  "theme_color": "#ffffff",
+  "background_color": "#ffffff",
+  "display": "standalone"
+}
